requirements.txt
# NVIDIA GPU monitoring
nvidia-ml-py3

# Graph generation
matplotlib
Pillow

# Logging
loguru
